is-lnu-service/src/test/java/org/lnu/is/extractor/addresstype/AddressTypeParametersExtractorTest.java
package org.lnu.is.extractor.addresstype;

import static org.junit.Assert.assertEquals;

import java.util.HashMap;
import java.util.Map;

import org.junit.Test;
import org.lnu.is.domain.addresstype.AddressType;

public class AddressTypeParametersExtractorTest {
	private AddressTypeParametersExtractor unit = new AddressTypeParametersExtractor();

	@Test
	public void testGetParameters() throws Exception {
		// Given
		String name = "AddressN";
		String abbrName = "AN";

		AddressType entity = new AddressType();
		entity.setName(name);
		entity.setAbbrName(abbrName);

		Map<String, Object> expected = new HashMap<String, Object>();
		expected.put("name", name);
		expected.put("abbrName", abbrName);

		// When
		Map<String, Object> actual = unit.getParameters(entity);

		// Then
		assertEquals(expected, actual);
	}

}
